requirements.txt
opencv-python
numpy
pytesseract
ffmpeg-python
tqdm
